+++ src/main/java/ru/hometask/dto/UpdateContractDto.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.Min;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,7 +16,7 @@ public class UpdateContractDto {
 
     private Long clientId;
 
-    private int amount;
+    private double amount;
 
     private LocalDateTime dateOfIssue;
